fix undefined behavior

+++ 3.ch/hashtable.c
@@ -22,7 +22,7 @@ typedef struct {
 
 // простая хэш функция
 size_t hash(const char *key, size_t size){
-	unsigned int hash = 5381;
+	size_t hash = 5381;
 	for(size_t i = 0; key[i] != '\0';i++){
 		if(isalnum(key[i])){//исключаем знаки препинания (не работает, не знаю почему) , думаю из-за кодировки , но не уверен
 			hash = hash * 33 << key[i];
